fix: 컴포넌트 점수 자료형 변경 int -> Long

+++ src/main/java/com/weflo/backend/domain/drone/domain/DroneComponent.java
@@ -19,7 +19,7 @@ public class DroneComponent {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "drone_component_id")
     private Long id;
-    private int point;
+    private Long point;
     private int brokenCount;
     @ManyToOne
     @JoinColumn(name = "component_id")
